Rename betty.console.rich to betty.rich
