+++ betty/rich/__init__.py
@@ -1,3 +1,19 @@
 """
 Rich integration.
 """
+
+from rich.style import Style
+from rich.theme import Theme
+
+
+class ConsoleTheme(Theme):
+    """
+    A Rich theme for Betty's console.
+    """
+
+    def __init__(self):
+        super().__init__(
+            {
+                "prompt": Style(color="bright_magenta", bold=True),
+            }
+        )
