+++ betty/rich/user.py
@@ -15,11 +15,11 @@ from typing_extensions import override
 
 from betty.assertion import Assertion
 from betty.console.progress import ConsoleProgress
-from betty.console.rich import ConsoleTheme
 from betty.locale.localizable import LocalizableLike
 from betty.locale.localized import ensure_localized
 from betty.progress import Progress
 from betty.progress.no_op import NoOpProgress
+from betty.rich import ConsoleTheme
 from betty.typing import Void, internal
 from betty.user import User, Verbosity
 from betty.user.logging import UserHandler
